requirements/base.txt
twisted==20.3.0 # fix for uvicorn/daphne error
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
uvicorn==0.11.8  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.0.10  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.42.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.11.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.5.0 # https://github.com/adamchainz/django-cors-headers
django-webpack-loader==0.6.0  # https://github.com/owais/django-webpack-loader

# [OWN_DD]

django-filter==2.4.0
django-lockdown==3.0.0
django-simple-history==2.12.0
djangochannelsrestframework==0.2.0
django-rest-auth==0.9.5

channels==3.0.2
channels-demultiplexer==1.0.1
channels-redis==3.2.0
channelsmultiplexer==0.0.3

python-dateutil==2.8.1
django-background-tasks==1.2.5

drf-yasg==1.20.0

# for UML visualization

# pydotplus==2.0.2
# django-extensions==3.1.1
